Moved common methods to base
