+++ app/src/main/java/com/drdisagree/iconify/ui/fragments/xposed/LockscreenWeather.kt
@@ -5,6 +5,7 @@ import com.drdisagree.iconify.R
 import com.drdisagree.iconify.common.Preferences.WEATHER_OWM_KEY
 import com.drdisagree.iconify.common.Preferences.WEATHER_PROVIDER
 import com.drdisagree.iconify.common.Preferences.WEATHER_SWITCH
+import com.drdisagree.iconify.common.Preferences.WEATHER_YANDEX_KEY
 import com.drdisagree.iconify.config.RPrefs
 import com.drdisagree.iconify.ui.activities.MainActivity
 import com.drdisagree.iconify.ui.base.WeatherPreferenceFragment
@@ -28,16 +29,6 @@ class LockscreenWeather : WeatherPreferenceFragment() {
         return WEATHER_SWITCH
     }
 
-    private fun showOwnKeyDialog() {
-        MaterialAlertDialogBuilder(requireContext())
-            .setTitle(appContextLocale.getString(R.string.weather_provider_owm_key_title))
-            .setMessage(appContextLocale.getString(R.string.weather_provider_owm_key_message))
-            .setCancelable(false)
-            .setPositiveButton(appContextLocale.getString(R.string.understood), null)
-            .create()
-            .show()
-    }
-
     override fun updateScreen(key: String?) {
         super.updateScreen(key)
 
@@ -47,14 +38,6 @@ class LockscreenWeather : WeatherPreferenceFragment() {
                     activityBinding = (requireActivity() as MainActivity).binding,
                     requiresSystemUiRestart = true
                 )
-            }
-
-            WEATHER_PROVIDER -> {
-                if (RPrefs.getString(WEATHER_PROVIDER) == "1" && RPrefs.getString(WEATHER_OWM_KEY)
-                        .isNullOrEmpty()
-                ) {
-                    showOwnKeyDialog()
-                }
             }
         }
     }
